+++ src/otx/backend/openvino/engine.py
@@ -204,7 +204,6 @@ class OVEngine(Engine):
                 f"datamodule.label_info={self.datamodule.label_info}"
             )
             raise ValueError(msg)
-
         metric_callable = metric(datamodule.label_info)
         with Progress() as progress:
             dataloader = datamodule.test_dataloader()
@@ -270,10 +269,7 @@ class OVEngine(Engine):
             ValueError: If input data is invalid or label information does not match.
             TypeError: If input data type is unsupported.
         """
-        from otx.backend.native.models.utils.xai_utils import (
-            process_saliency_maps_in_pred_entity,
-            set_crop_padded_map_flag,
-        )
+        from otx.backend.native.models.utils.xai_utils import process_saliency_maps_in_pred_entity
 
         model = self._update_checkpoint(checkpoint)
         if isinstance(data, (str, os.PathLike)):
@@ -329,7 +325,6 @@ class OVEngine(Engine):
         if explain and isinstance(datamodule, OTXDataModule):
             if explain_config is None:
                 explain_config = ExplainConfig()
-            explain_config = set_crop_padded_map_flag(explain_config, datamodule)
             predict_result = process_saliency_maps_in_pred_entity(predict_result, explain_config, datamodule.label_info)
 
         return predict_result