+++ src/otx/backend/openvino/models/instance_segmentation.py
@@ -11,9 +11,9 @@ import torch
 from model_api.tilers import InstanceSegmentationTiler
 from torchvision import tv_tensors
 
-from otx.backend.native.models.instance_segmentation.utils.structures.mask.mask_util import encode_rle, polygon_to_rle
 from otx.backend.openvino.models.base import OVModel
 from otx.data.entity.torch import OTXDataBatch, OTXPredBatch
+from otx.data.utils.structures.mask.mask_util import encode_rle, polygon_to_rle
 from otx.metrics import MetricInput
 from otx.metrics.mean_ap import MaskRLEMeanAPFMeasureCallable
 from otx.types.label import LabelInfo