+++ tests/unit/backend/native/tools/adaptive_bs/test_adaptive_bs_api.py
@@ -17,6 +17,7 @@ from otx.backend.native.tools.adaptive_bs.runner import (
     _train_model,
     adapt_batch_size,
 )
+from otx.utils.device import is_xpu_available
 
 
 @pytest.fixture()
@@ -180,15 +181,18 @@ def test_adapt_batch_size_dist_sub_proc(
     assert int(mock_os.environ["ADAPTIVE_BS_FOR_DIST"]) == cur_bs
 
 
-def test_adapt_batch_size_no_accelerator(
+def test_adapt_batch_size_cpu(
     mock_is_cuda_available,
     mock_is_xpu_available,
     mock_engine,
     train_args,
+    mocker,
 ):
     mock_is_cuda_available.return_value = False
-    with pytest.raises(RuntimeError, match="Adaptive batch size supports CUDA or XPU."):
-        adapt_batch_size(mock_engine, **train_args)
+    mock_is_xpu_available.return_value = False
+    mock_logger = mocker.patch("otx.backend.native.tools.adaptive_bs.runner.logger")
+    adapt_batch_size(mock_engine, **train_args)
+    mock_logger.warning.assert_called_once_with("Adaptive batch size supports only CUDA or XPU.")
 
 
 def test_adjust_train_args(train_args):
@@ -260,7 +264,7 @@ class TestBatchSizeFinder:
         # check steps_per_trial is set well
         assert mock_trainer.limit_val_batches == steps_per_trial
         assert mock_trainer.fit_loop.epoch_loop.max_steps == -1
-        assert mock_trainer.fit_loop.max_epochs == 1
+        assert mock_trainer.fit_loop.max_epochs == 1 if not is_xpu_available() else 2
         assert mock_trainer.limit_train_batches == steps_per_trial
         # check active_loop is run
         assert mock_active_loop.restarting is False
@@ -278,7 +282,7 @@ class TestBatchSizeFinder:
         # check steps_per_trial is set well
         assert mock_trainer.limit_val_batches == 0
         assert mock_trainer.fit_loop.epoch_loop.max_steps == -1
-        assert mock_trainer.fit_loop.max_epochs == 1
+        assert mock_trainer.fit_loop.max_epochs == 1 if not is_xpu_available() else 2
         assert mock_trainer.limit_train_batches == steps_per_trial
         # check active_loop is run
         assert mock_active_loop.restarting is False